app/requirements.txt
aiohttp==3.8.4
fastapi==0.97.0
python-dotenv==1.0.0
uvicorn==0.22.0
